requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Django==2.2.3
django-extensions==2.1.7
importlib-metadata==0.17
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.0
pytest==4.6.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
wcwidth==0.1.7
zipp==0.5.1
django-crispy-forms==1.7.2
psycopg2==2.8.3

arrow==0.14.2
